Add fake _latest target for C++ KeyMint
The `keystore2` service depends on the latest version of the KeyMint
HAL, but also includes a copy of the C++ KeyMint implementation code (to
support software-emulation of some features).  The C++ KeyMint code is
not going to support KeyMint versions above v4, so this will lead to
multiple-version errors when the KeyMint AIDL moves to v5.

To work around this, add a variant of the C++ target that claims to
support the latest KeyMint HAL, even though the code actually only
supports v4.  This target should not be used for an actual KeyMint
implementation (as it lies about the version supported and will not pass
VTS), so visibility-limit the new target to keystore2's compatibility
layer.

Test: TreeHugger
Flag: EXEMPT build refactor
Bug: 379008955
Bug: 395069628
Change-Id: I873bfcc3df4d03a4aa58c424025e3420bf1abf15